requirements.txt
timm
fastcore
tqdm
wandb
torch_pruning>=1.4.1
tqdm
huggingface_hub==0.15.1
importlib_metadata
fastcore
fvcore
ptflops